Initial decouple Game from its Loading/Saving features
Create a class which will handle saving/loading and creations
of Games, that class can be extended.
