+++ src/Game.h
@@ -25,25 +25,8 @@ namespace Graphics {
 	class Renderer;
 }
 
-struct CannotSaveCurrentGameState {};
-struct CannotSaveInHyperspace : public CannotSaveCurrentGameState {};
-struct CannotSaveDeadPlayer : public CannotSaveCurrentGameState {};
-struct InvalidGameStartLocation {
-	std::string error;
-	InvalidGameStartLocation(const std::string &error_) :
-		error(error_) {}
-};
-
 class Game {
 public:
-	static Json LoadGameToJson(const std::string &filename);
-	// LoadGame and SaveGame throw exceptions on failure
-	static Game *LoadGame(const std::string &filename);
-	static bool CanLoadGame(const std::string &filename);
-	// XXX game arg should be const, and this should probably be a member function
-	// (or LoadGame/SaveGame should be somewhere else entirely)
-	static void SaveGame(const std::string &filename, Game *game);
-
 	// start docked in station referenced by path or nearby to body if it is no station
 	Game(const SystemPath &path, const double startDateTime = 0.0);
